backend/pyproject.toml
[project]
name = "dociq-backend"
version = "0.1.0"
description = "DocIQ backend modular monolith"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.2",
  "asyncpg>=0.29.0",
  "boto3>=1.34.162",
  "celery[redis]>=5.4.0",
  "cryptography>=43.0.1",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.2",
  "itsdangerous>=2.2.0",
  "numpy>=2.1.0",
  "opencv-python-headless>=4.10.0.84",
  "orjson>=3.10.7",
  "passlib[bcrypt]>=1.7.4",
  "pillow>=10.4.0",
  "prometheus-fastapi-instrumentator>=7.0.0",
  "psycopg[binary]>=3.2.1",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "pymupdf>=1.24.9",
  "pyjwt>=2.9.0",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.9",
  "qdrant-client>=1.10.1",
  "redis>=5.0.8",
  "sqlalchemy>=2.0.32",
  "structlog>=24.4.0",
  "tenacity>=9.0.0",
  "typer>=0.12.3",
  "uuid6>=2024.7.10",
  "uvicorn[standard]>=0.30.6"
]

[project.optional-dependencies]
dev = [
  "coverage>=7.6.1",
  "faker>=26.2.0",
  "freezegun>=1.5.1",
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "respx>=0.21.1",
  "ruff>=0.6.1"
]

[tool.pytest.ini_options]
testpaths = ["../tests/backend", "../tests/ai_pipeline"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "S", "ASYNC"]
ignore = ["B008"]

[build-system]
requires = ["setuptools>=75.1.0", "wheel"]
build-backend = "setuptools.build_meta"

